Move read_mgmt_files to npu_mgmt module
